requirements.txt
python-dotenv
langchain
pandas
langchain-openai
langchain_core
streamlit>=1.25
openai>=1.0.0
